places query with filter
