+++ pages/map.tsx
@@ -1,62 +1,209 @@
+import { ApolloQueryResult } from '@apollo/client';
 import { Layout } from '@components/Layout';
-import { useMapPostsQuery } from '@graphql/geo.graphql';
-import { useIsLoggedQuery } from '@graphql/user.graphql';
-import { Search } from 'components/MainPage';
-import { MapGL } from 'components/Map';
-import React, { useEffect, useState } from 'react';
+import { PlacesQuery, usePlacesQuery } from '@graphql/geo.graphql';
+import Image from 'next/image';
+import { NextRouter, useRouter } from 'next/router';
+import React, { useEffect, useRef, useState } from 'react';
+import ReactMapGL, { ExtraState, MapRef, Marker } from 'react-map-gl';
 
-const MapPage: React.FC = () => {
-  const { data, loading, error } = useIsLoggedQuery();
-  const [bounds, setBounds] = useState<{
-    maxLatitude: number;
-    minLatitude: number;
-    maxLongitude: number;
-    minLongitude: number;
-  }>({
-    maxLatitude: 0,
-    minLatitude: 0,
-    maxLongitude: 0,
-    minLongitude: 0,
+type Viewport = {
+  latitude: number;
+  longitude: number;
+  zoom: number;
+};
+
+type Bounds = {
+  maxLatitude: number;
+  minLatitude: number;
+  maxLongitude: number;
+  minLongitude: number;
+};
+
+function PropsToUrl({
+  pathname,
+  router,
+  query,
+}: {
+  pathname: string;
+  router: NextRouter;
+  query: any;
+}): void {
+  router.replace({
+    pathname,
+    query,
   });
-  const [points, setPoints] = useState<Array<any>>([]);
+}
 
-  const posts = useMapPostsQuery({ variables: bounds });
+const Map: React.FC = () => {
+  const router = useRouter();
 
-  const [searchCoordinates, setSearchCoordinates] = useState({
-    lat: 50,
-    lng: 15,
+  const { data, loading, error, refetch } = usePlacesQuery({
+    variables: {
+      input: {
+        filter: { take: 1 },
+      },
+    },
   });
 
+  const [timeLimitation, setTimeLimitation] = useState<[number, number]>([
+    1000,
+    new Date().getTime(),
+  ]);
+
+  const [sidebarPlace, setSidebarPlace] = useState<number | null>(null);
+
+  // map viewport
+  const [mapViewport, setMapViewport] = useState<Viewport>({
+    latitude: 37.7577,
+    longitude: -122.4376,
+    zoom: 3.5,
+  });
+
+  // on load change map viewport coordinates according to url parameter
   useEffect(() => {
-    // filter points from query by point IDs
-    const tmp = posts.data?.mapPosts
-      ? posts.data.mapPosts.filter(
-          (x) => points.find((y) => y.id === x?.id) === undefined
-        )
-      : [];
+    /*
+    function paramToNum(parameter: any, defaultValue: any) {
+      return typeof parameter == 'string'
+        ? parseFloat(parameter.substring(0, 9))
+        : defaultValue;
+    }
+    
+    setMapViewport({
+      ...mapViewport,
+      longitude: paramToNum(router.query?.lng, 15),
+      latitude: paramToNum(router.query?.lat, 50),
+      zoom: paramToNum(router.query?.zoom, 14),
+      // @ts-ignore
+      transitionDuration: 500,
+      transitionInterpolator: new FlyToInterpolator(),
+    })
+    */
+    // @ts-ignore
+    // setSidebarPlace(paramToNum(router.query?.place, null));
+  }, []);
 
-    // concat already existing points with new ones
-    setPoints([...tmp, ...points]);
-  }, [posts.data]);
-
-  useEffect(() => {
-    posts.refetch(bounds);
-  }, [bounds]);
-
-  if (loading) return <div></div>;
-  if (error) return <div></div>;
   return (
     <Layout title="map | hiStories">
-      <div className="w-full" style={{ height: 'calc(100vh - 14px)' }}>
-        <MapGL
-          searchCoordinates={searchCoordinates}
-          setBounds={setBounds}
-          oldPoints={points}
-        />
-      </div>
-      <Search setSearchCoordinates={setSearchCoordinates} />
+      <section className="flex w-full h-screen">
+        <div id="map" className="w-full h-full">
+          <MapGL
+            viewport={mapViewport}
+            setViewport={setMapViewport}
+            data={data}
+            refetch={refetch}
+            setSidebar={setSidebarPlace}
+            sidebar={sidebarPlace}
+          />
+        </div>
+        <div id="sidebar" className="w-full h-full">
+          {JSON.stringify(data?.places.length)}
+        </div>
+      </section>
     </Layout>
   );
 };
 
-export default MapPage;
+function ConvertBounds(bounds: {
+  _ne: { lat: number; lng: number };
+  _sw: { lat: number; lng: number };
+}): Bounds {
+  return {
+    maxLatitude: bounds._ne.lat,
+    minLatitude: bounds._sw.lat,
+    maxLongitude: bounds._ne.lng,
+    minLongitude: bounds._sw.lng,
+  };
+}
+
+type MapGLProps = {
+  viewport: Viewport;
+  setViewport: React.Dispatch<React.SetStateAction<Viewport>>;
+  setSidebar: React.Dispatch<React.SetStateAction<number | null>>;
+  data: PlacesQuery | undefined;
+  refetch: (args: any) => Promise<ApolloQueryResult<PlacesQuery>>;
+  sidebar: number | null;
+};
+
+const MapGL: React.FC<MapGLProps> = ({
+  viewport,
+  setViewport,
+  data,
+  refetch,
+  setSidebar,
+  sidebar,
+}) => {
+  const mapRef = useRef<MapRef | null>(null);
+
+  return (
+    <ReactMapGL
+      {...viewport}
+      width="100%"
+      height="100%"
+      onViewportChange={(viewport: any) => setViewport(viewport)}
+      mapboxApiAccessToken={process.env.NEXT_PUBLIC_MAPBOX_TOKEN} // MAPBOX API ACCESS TOKEN
+      mapStyle="mapbox://styles/mapbox/streets-v11" // MAPBOX STYLE
+      ref={(instance) => (mapRef.current = instance)}
+      onLoad={async () => {
+        // if map is rendered get bounds and refetch
+        if (mapRef.current)
+          await refetch(
+            // get bounds of map
+            ConvertBounds(mapRef.current.getMap().getBounds())
+          );
+      }}
+      onInteractionStateChange={async (state: ExtraState) => {
+        // when map state changes (dragging, zooming, rotating, etc.)
+        if (!state.isDragging && mapRef.current) {
+          // refetch data
+          await refetch({
+            input: {
+              filter:
+                // get bounds of map
+                ConvertBounds(mapRef.current.getMap().getBounds()),
+            },
+          });
+
+          console.log(data);
+        }
+      }}
+    >
+      {data?.places
+        .filter((place) => place.preview || place.icon)
+        .map((place) => {
+          return (
+            <Marker
+              key={place.id}
+              latitude={place.latitude}
+              longitude={place.longitude}
+            >
+              <div
+                onClick={() => setSidebar(place.id)}
+                className="cursor-pointer -translate-y-1/2 -translate-x-1/2"
+              >
+                {place.icon ? (
+                  <Image
+                    src={place.icon}
+                    width={90}
+                    height={90}
+                    objectFit="contain"
+                    alt="Picture on map"
+                  />
+                ) : (
+                  <Image
+                    // @ts-ignore
+                    src={place.preview[0]}
+                    width={60}
+                    height={60}
+                    className="object-cover rounded-full"
+                    alt="Picture on map"
+                  />
+                )}
+              </div>
+            </Marker>
+          );
+        })}
+    </ReactMapGL>
+  );
+};
+
+export default Map;
